requirements.txt
# Core Dependencies
python-dotenv>=1.0.0  # For loading environment variables
openai>=1.3.0         # OpenAI API client
langchain>=0.0.267    # LangChain framework
neo4j>=5.8.1          # Neo4j database driver
spacy>=3.5.3          # For NLP tasks
pandas>=2.0.0         # Data manipulation
numpy>=1.24.0         # Numerical operations
requests>=2.28.0      # HTTP requests
rdflib>=6.3.2         # RDF processing
pydantic>=2.0.0       # Data validation

# NLP and Text Processing
nltk>=3.8.1           # Natural language toolkit
fuzzywuzzy>=0.18.0    # Fuzzy string matching
python-Levenshtein>=0.21.0  # Faster implementation for fuzzywuzzy
chardet>=5.1.0        # Character encoding detection

# Visualization and Analysis
matplotlib>=3.7.1     # Plotting and visualization
seaborn>=0.12.2       # Statistical data visualization
networkx>=3.1         # Network analysis

# Jupyter and Development Tools
jupyter>=1.0.0        # Jupyter notebook
ipython>=8.12.0       # Interactive Python shell
pytest>=7.3.1         # Testing framework

# Optional: Install spaCy language model
# Run after installing requirements:
# python -m spacy download en_core_web_sm